requirements.txt
PyYAML==6.0
pandas==1.5.1